fix issues and refactor test
