+++ src/test/movieSlice.test.js
@@ -1,12 +1,12 @@
-import moviesSlice, { fetchMovies } from '../data/moviesSlice'
+import moviesSlice, { fetchMovies } from '../store/moviesSlice'
 import { moviesMock } from './movies.mocks'
 
 describe('MovieSlice test', () => {
-    
+
     it('should set loading true while action is pending', () => {
         const action = {type: fetchMovies.pending};
         const initialState = moviesSlice.reducer(
-        { 
+        {
             movies: [], fetchStatus: '',
         }, action);
         expect(action).toEqual({type: fetchMovies.pending})
@@ -14,11 +14,11 @@ describe('MovieSlice test', () => {
 
     it('should return payload when action is fulfilled', () => {
         const action = {
-            type: fetchMovies.fulfilled, 
+            type: fetchMovies.fulfilled,
             payload: moviesMock
         };
         const initialState = moviesSlice.reducer(
-        { 
+        {
             movies: [], fetchStatus: '',
         }, action);
         expect(action.payload).toBeTruthy()
@@ -27,7 +27,7 @@ describe('MovieSlice test', () => {
     it('should set error when action is rejected', () => {
         const action = {type: fetchMovies.rejected};
         const initialState = moviesSlice.reducer(
-        { 
+        {
             movies: [], fetchStatus: '',
         }, action);
         expect(action).toEqual({type: fetchMovies.rejected})
