+++ src/test/starredSlice.test.js
@@ -1,4 +1,4 @@
-import starredSlice from '../data/starredSlice'
+import starredReducer, {starMovie, unstarMovie, clearAllStarred} from '../store/starredSlice'
 import { moviesMock } from './movies.mocks'
 
 describe('starredSlice test', () => {
@@ -8,28 +8,28 @@ describe('starredSlice test', () => {
     it('should set an initial state', () => {
         const initialState = state
         const action = { type: '' }
-        const result = starredSlice.reducer(initialState, action)
+        const result = starredReducer(initialState, action)
         expect(result).toEqual({ starredMovies: []})
-      })    
+      })
 
       it('should add movie to starred', () => {
         const initialState = { ...state, starredMovies: [] }
-        const action = starredSlice.actions.starMovie(moviesMock[0])
-        const result = starredSlice.reducer(initialState, action)
+        const action = starMovie(moviesMock[0])
+        const result = starredReducer(initialState, action)
         expect(result.starredMovies[0]).toBe(moviesMock[0])
       })
 
       it('should remove movie from starred', () => {
         const initialState = { ...state, starredMovies: moviesMock }
-        const action = starredSlice.actions.unstarMovie(moviesMock[0])
-        const result = starredSlice.reducer(initialState, action)
+        const action = unstarMovie(moviesMock[0])
+        const result = starredReducer(initialState, action)
         expect(result.starredMovies[0]).toBe(moviesMock[1])
       })
 
       it('should remove all movies', () => {
         const initialState = { ...state, starredMovies: moviesMock }
-        const action = starredSlice.actions.clearAllStarred(state)
-        const result = starredSlice.reducer(initialState, action)
+        const action = clearAllStarred(state)
+        const result = starredReducer(initialState, action)
         expect(Object.keys(result.starredMovies).length).toEqual(0)
       })
 })
